+++ framework/CHANGELOG.md
@@ -3,6 +3,7 @@ Yii Framework 2 Change Log
 
 2.0.12 under development
 --------------------------
+- Enh #14145: Add an ImmutableValidator to stop model attribute being modified
 
 - Bug #13842: Fixed ambiguous table SQL error while using `yii\validators\ExistValidator` and `yii\validators\UniqueValidator` (vladis84, samdark)
 - Bug #5442: Fixed problem on load fixture dependencies with database related tests (leandrogehlen)